+++ test/test-helper.el
@@ -1,5 +1,5 @@
 (require 'f)
-(require 'web-server)
+(require 'simple-httpd)
 
 ;; Do not truncate the backtrace.  This makes ERT easy to debug.
 (setq ert-batch-backtrace-right-margin 256)
@@ -11,15 +11,3 @@
 (defvar fixture-root
   (f-join test-path "fixtures")
   "Path to fixture root directory.")
-
-(defun make-static-handler (source)
-  (lambda (request)
-    (with-slots (process headers) request
-      (let ((path (substring (cdr (assoc :GET headers)) 1)))
-        (if (ws-in-directory-p source path)
-            (if (file-directory-p path)
-                (ws-send-directory-list process
-                                        (f-join source path)
-                                        "^[^\.]")
-              (ws-send-file process (f-join source path)))
-          (ws-send-404 process))))))